requirements.txt
requests
python-swiftclient
pyyaml